test_requirements.txt
pytest==6.2.5
pytest-asyncio==0.16.0
factory-boy==3.2.1
requests==2.26.0